// plugins/client/src/configuration.validator.ts
// Generated by ts-to-zod
import { z } from "zod"

export const generatorInputSchema = z.object({
    raw: z.string().optional(),
    base: z.string().optional(),
    file: z.string().optional(),
    include: z.union([z.string(), z.array(z.string())]).optional(),
    exclude: z.union([z.string(), z.array(z.string())]).optional(),
    project: z.union([z.string(), z.array(z.string())]).optional(),
})

export const customTypesMapSchema = z.object({
    String: z.string().optional(),
    Guid: z.string().optional(),
    Uri: z.string().optional(),
    Boolean: z.string().optional(),
    UInt8: z.string().optional(),
    Int8: z.string().optional(),
    Int16: z.string().optional(),
    UInt16: z.string().optional(),
    Int32: z.string().optional(),
    UInt32: z.string().optional(),
    Int64: z.string().optional(),
    UInt64: z.string().optional(),
    Float32: z.string().optional(),
    Float64: z.string().optional(),
    DateOnly: z.string().optional(),
    TimeOnly: z.string().optional(),
    DateTimeOffset: z.string().optional(),
    TimeSpan: z.string().optional(),
})

export const clientGeneratorPluginConfigurationSchema = z.object({
    input: generatorInputSchema,
    customTypes: customTypesMapSchema.optional(),
    nameTransform: z.function().args(z.string()).returns(z.string()).optional(),
})